+++ bindgen-tests/tests/tests.rs
@@ -672,22 +672,6 @@ fn dump_preprocessed_input() {
     );
 }
 
-#[test]
-fn allowlist_warnings() {
-    let header = concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/tests/headers/allowlist_warnings.h"
-    );
-
-    let bindings = builder()
-        .header(header)
-        .allowlist_function("doesnt_match_anything")
-        .generate()
-        .expect("unable to generate bindings");
-
-    assert_eq!(1, bindings.warnings().len());
-}
-
 fn build_flags_output_helper(builder: &bindgen::Builder) {
     let mut command_line_flags = builder.command_line_flags();
     command_line_flags.insert(0, "bindgen".to_string());